+++ core/src/AzureMcp.Core/Areas/Server/Commands/ToolLoading/CompositeToolLoader.cs
@@ -1,10 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using System.Threading;
 using Microsoft.Extensions.Logging;
 using ModelContextProtocol.Protocol;
-using ModelContextProtocol.Server;
 
 namespace AzureMcp.Core.Areas.Server.Commands.ToolLoading;